ttf2pk(2): Use binary mode for output files Change function params to int to avoid MSVC warnings without casting Drop XCDECL (from W32TeX)
git-svn-id: svn://tug.org/texlive/trunk/Build/source@27080 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/ttf2pk/pklib.c
@@ -652,7 +652,7 @@ pk_bm_cvt(void)
 
 
 static void
-putshort(short w)
+putshort(int w)
 {
   putc(w >> 8, pk_file);
   putc(w, pk_file);
